+++ src/components/common/utils/StarTotal.js
@@ -3,7 +3,7 @@ import numeral from 'numeral'
 
 import StarIcon from './StarIcon'
 
-const StarTotal = ({ value, icon = true }) => {
+const StarTotal = ({ value }) => {
   const digits = value > 1000 && value < 10000 ? '0.0' : '0'
   return (
     <span>